refactor transformDate and transformSrvNodes
